Code-style updates
Added husky and lint-staged;
Move src → lib
