+++ .editorconfig
@@ -3,7 +3,7 @@ root = true
 # Unix-style newlines with a newline ending every file
 [*]
 end_of_line = lf
-insert_final_newline = true
+insert_final_newline = false
 indent_style = space
 indent_size = 2
 trim_trailing_whitespace = true
