+++ .eslintrc
@@ -1,10 +1,14 @@
 {
+  "parserOptions": {
+    "ecmaVersion": 6
+  },
   "plugins": [],
   "extends": [
     "eslint:recommended"
   ],
   "rules": {
     "comma-spacing": ["warn"],
+    "eol-last": ["warn", "never"],
     "indent": ["warn", 2, {"SwitchCase": 1}],
     "key-spacing": ["warn", {
       "beforeColon": false,
@@ -12,10 +16,11 @@
     }],
     "keyword-spacing": ["warn", {"after": true}],
     "no-extra-boolean-cast": "off",
+    "no-multiple-empty-lines": ["warn", {"max": 1}],
     "no-multi-spaces": "warn",
     "no-spaced-func": "warn",
     "no-unused-vars": "warn",
-    "object-curly-spacing": ["warn", "never"],
+    "object-curly-spacing": ["warn", "always"],
     "semi": ["warn", "always"],
     "space-before-blocks": ["warn", "always"],
     "space-before-function-paren": ["warn", "never"],
@@ -28,8 +33,6 @@
     }]
   },
   "env": {
-    "browser": true,
-    "commonjs": true,
     "node": true
   }
 }
