+++ examples/info.js
@@ -1,10 +1,9 @@
-import info from '../src/info';
+import info from '../lib/info';
 
 const API_TOKEN = '';
 
-info(API_TOKEN, ({total_space, used_space}) => {
+info(API_TOKEN, ({ total_space, used_space }) =>
   console.log(`
     Total space: ${Math.round(total_space / 1000000000)}GB
     Free space: ${Math.round((total_space - used_space) / 1000000)}MB
-  `);
-});
+  `));
