+++ examples/list.js
@@ -1,5 +1,5 @@
-import list from '../src/list';
+import list from '../lib/list';
 
 const API_TOKEN = '';
 
-list(API_TOKEN, {}, ({items}) => items.forEach(({name, size}) => console.log('%s: %dB', name, size)));
+list(API_TOKEN, {}, ({ items }) => items.forEach(({ name, size }) => console.log('%s: %dB', name, size)));
