+++ examples/meta.js
@@ -1,7 +1,7 @@
-import meta from '../src/meta';
+import meta from '../lib/meta';
 
-const API_TOKEN = 'AQAAAAAbX24dAAQB-z80p6HY-02ts2aztofaFeQ';
+const API_TOKEN = '';
 
-meta.add(API_TOKEN, 'disk:/Зима.jpg', {my_field: 'my_value'}, null, console.error);
+meta.add(API_TOKEN, 'disk:/Зима.jpg', { my_field: 'my_value' }, null, console.error);
 
 meta.get(API_TOKEN, 'disk:/Зима.jpg', {}, console.log);
