+++ lib/download.js
@@ -1,6 +1,6 @@
 const request = require('./request');
 
-const {API_DOWNLOAD_LINK_URL} = require('./constants');
+const { API_DOWNLOAD_LINK_URL } = require('./constants');
 
 /**
  * Get download link for the file
@@ -10,15 +10,14 @@ const {API_DOWNLOAD_LINK_URL} = require('./constants');
  * @param {function} [success] Success callback
  * @param {function} [error] Error callback
  */
-const link = (token, path, success, error) => {
+const link = (token, path, success, error) =>
   request.get({
     url: API_DOWNLOAD_LINK_URL,
     token: token,
     query: {
-      path: path
+      path
     }
   }, success, error);
-};
 
 module.exports = {
   link
