+++ lib/info.js
@@ -1,5 +1,5 @@
 const request = require('./request');
-const {API_DISK_URL} = require('./constants');
+const { API_DISK_URL } = require('./constants');
 
 /**
  * Getting info about the disk info
@@ -8,11 +8,10 @@ const {API_DISK_URL} = require('./constants');
  * @param {function} [success] Success callback
  * @param {function} [error] Error callback
  */
-const info = (token, success, error) => {
+const info = (token, success, error) =>
   request.get({
     url: API_DISK_URL,
     token: token
   }, success, error);
-};
 
 module.exports = info;
