+++ lib/list.js
@@ -1,6 +1,6 @@
 const request = require('./request');
 
-const {API_FILES_URL} = require('./constants');
+const { API_FILES_URL } = require('./constants');
 
 /**
  * Getting flat list of files
@@ -15,12 +15,11 @@ const {API_FILES_URL} = require('./constants');
  * @param {function} [success] Success callback
  * @param {function} [error] Error callback
  */
-const list = (token, options, success, error) => {
+const list = (token, options, success, error) =>
   request.get({
     url: API_FILES_URL,
     token: token,
     query: options
   }, success, error);
-};
 
 module.exports = list;
