+++ lib/meta.js
@@ -1,6 +1,6 @@
 const request = require('./request');
 
-const {API_RESOURCES_URL} = require('./constants');
+const { API_RESOURCES_URL } = require('./constants');
 
 /**
  * Request meta-info for the resource (file or folder)
@@ -16,13 +16,12 @@ const {API_RESOURCES_URL} = require('./constants');
  * @param {function} [success] Success callback
  * @param {function} [error] Error callback
  */
-const get = (token, path, options = {}, success, error) => {
+const get = (token, path, options = {}, success, error) =>
   request.get({
     url: API_RESOURCES_URL,
     token: token,
-    query: Object.assign({path}, options)
+    query: Object.assign({ path }, options)
   }, success, error);
-};
 
 /**
  * Append meta-info to the resource (file or folder)
@@ -33,18 +32,17 @@ const get = (token, path, options = {}, success, error) => {
  * @param {function} [success] Success callback
  * @param {function} [error] Error callback
  */
-const add = (token, path, props, success, error) => {
+const add = (token, path, props, success, error) =>
   request.patch({
     url: API_RESOURCES_URL,
     token: token,
     query: {
-      path: path
+      path
     },
     data: {
       custom_properties: props
     },
   }, success, error);
-};
 
 module.exports = {
   get,
