+++ lib/recent.js
@@ -1,6 +1,6 @@
 const request = require('./request');
 
-const {API_RECENT_FILES_URL} = require('./constants');
+const { API_RECENT_FILES_URL } = require('./constants');
 
 /**
  * Get list of last uploaded files
@@ -14,12 +14,11 @@ const {API_RECENT_FILES_URL} = require('./constants');
  * @param {function} [success] Success callback
  * @param {function} [error] Error callback
  */
-const recent = (token, options, success, error) => {
+const recent = (token, options, success, error) =>
   request.get({
     url: API_RECENT_FILES_URL,
     token: token,
     query: options
   }, success, error);
-};
 
 module.exports = recent;
