+++ lib/request.js
@@ -1,6 +1,6 @@
 const https = require('https');
-const {parse: urlParse} = require('url');
-const {stringify: queryStringify} = require('querystring');
+const { parse: urlParse } = require('url');
+const { stringify: queryStringify } = require('querystring');
 
 const defaultRequestParams = {
   method: 'GET',
@@ -36,7 +36,7 @@ const request = (options, success, error) => {
     ),
     (res) => {
       let data = '';
-      const {statusCode} = res;
+      const { statusCode } = res;
 
       res.on('data', (chunk) => data += chunk);
       res.on('end', () => {
@@ -68,15 +68,15 @@ const request = (options, success, error) => {
 module.exports = {
   request: request,
   get: function(options, ...callbacks) {
-    this.request(Object.assign(options, {method: 'GET'}), ...callbacks);
+    this.request(Object.assign(options, { method: 'GET' }), ...callbacks);
   },
   post: function(options, ...callbacks) {
-    this.request(Object.assign(options, {method: 'POST'}), ...callbacks);
+    this.request(Object.assign(options, { method: 'POST' }), ...callbacks);
   },
   put: function(options, ...callbacks) {
-    this.request(Object.assign(options, {method: 'PUT'}), ...callbacks);
+    this.request(Object.assign(options, { method: 'PUT' }), ...callbacks);
   },
   patch: function(options, ...callbacks) {
-    this.request(Object.assign(options, {method: 'PATCH'}), ...callbacks)
+    this.request(Object.assign(options, { method: 'PATCH' }), ...callbacks);
   }
 };
