+++ lib/upload.js
@@ -1,6 +1,6 @@
 const request = require('./request');
 
-const {API_UPLOAD_LINK_URL} = require('./constants');
+const { API_UPLOAD_LINK_URL } = require('./constants');
 
 /**
  * Request for thee upload link
@@ -11,7 +11,7 @@ const {API_UPLOAD_LINK_URL} = require('./constants');
  * @param {function} [success] Success callback
  * @param {function} [error] Error callback
  */
-const link = (token, path, overwrite = false, success, error) => {
+const link = (token, path, overwrite = false, success, error) =>
   request.get({
     url: API_UPLOAD_LINK_URL,
     token: token,
@@ -20,7 +20,6 @@ const link = (token, path, overwrite = false, success, error) => {
       overwrite: overwrite
     }
   }, success, error);
-};
 
 module.exports = {
   link
