+++ package.json
@@ -12,11 +12,11 @@
   "devDependencies": {
     "ava": "^0.18.0",
     "babel-cli": "^6.23.0",
-    "babel-plugin-transform-es2015-modules-commonjs": "^6.23.0",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-es2015": "^6.22.0",
+    "babel-preset-env": "^1.5.2",
     "codecov": "^2.1.0",
     "follow-redirects": "^1.2.1",
+    "husky": "^0.13.4",
+    "lint-staged": "^4.0.0",
     "nyc": "^10.2.0",
     "sinon": "^1.17.7"
   },
@@ -27,50 +27,11 @@
   "directories": {
     "test": "tests/*.spec.js"
   },
-  "dependencies": {
-    "balanced-match": "^0.4.2",
-    "brace-expansion": "^1.1.6",
-    "browser-stdout": "^1.3.0",
-    "commander": "^2.9.0",
-    "concat-map": "^0.0.1",
-    "debug": "^2.2.0",
-    "diff": "^1.4.0",
-    "escape-string-regexp": "^1.0.5",
-    "formatio": "^1.1.1",
-    "fs.realpath": "^1.0.0",
-    "glob": "^7.0.5",
-    "graceful-readlink": "^1.0.1",
-    "growl": "^1.9.2",
-    "has-flag": "^1.0.0",
-    "inflight": "^1.0.6",
-    "inherits": "^2.0.1",
-    "json3": "^3.3.2",
-    "lodash._baseassign": "^3.2.0",
-    "lodash._basecopy": "^3.0.1",
-    "lodash._basecreate": "^3.0.3",
-    "lodash._getnative": "^3.9.1",
-    "lodash._isiterateecall": "^3.0.9",
-    "lodash.create": "^3.1.1",
-    "lodash.isarguments": "^3.1.0",
-    "lodash.isarray": "^3.0.4",
-    "lodash.keys": "^3.1.2",
-    "lolex": "^1.3.2",
-    "minimatch": "^3.0.3",
-    "minimist": "^0.0.8",
-    "mkdirp": "^0.5.1",
-    "ms": "^0.7.1",
-    "once": "^1.4.0",
-    "path-is-absolute": "^1.0.1",
-    "samsam": "^1.1.2",
-    "sinon": "^1.17.7",
-    "supports-color": "^3.1.2",
-    "util": "^0.10.3",
-    "wrappy": "^1.0.2"
-  },
   "scripts": {
     "example": "babel-node",
-    "test": "nyc ava tests/*.test.js",
-    "test:watch": "ava --watch tests/*.test.js",
+    "precommit": "lint-staged",
+    "test": "nyc ava",
+    "test:watch": "ava --watch",
     "test:debug": "node --inspect node_modules/ava/profile.js",
     "coverage": "nyc report --reporter=lcov > coverage.lcov && codecov --token=c6135c0d-58d0-47ff-8cca-7c38f31f637d"
   },
@@ -80,8 +41,14 @@
     "api"
   ],
   "ava": {
+    "files": [
+      "tests/*.spec.js"
+    ],
     "require": [
       "babel-core"
     ]
+  },
+  "lint-staged": {
+    "+(examples|lib|tests)/*.js": ["eslint --fix", "git add"]
   }
 }
