+++ tests/recent.spec.js
@@ -1,11 +1,11 @@
-import {mock} from 'sinon';
+import { mock } from 'sinon';
 import test from 'ava';
 
-import request from '../src/request';
-import recent from '../src/recent';
+import request from '../lib/request';
+import recent from '../lib/recent';
 
-import {API_TOKEN} from './constants';
-import {API_RECENT_FILES_URL} from '../src/constants';
+import { API_TOKEN } from './constants';
+import { API_RECENT_FILES_URL } from '../lib/constants';
 
 const options = {
   limit: 13,
